update order and cart

+++ demo/src/main/java/com/example/demo/service/CartService.java
@@ -12,15 +12,15 @@ public interface CartService {
     public Cart addProductToCart(Integer customerId, Long productId)
             throws CartException, CustomerException, ProductException;
 
-    public Cart removeProductFromCart(Integer customerId, Integer productId)
+    public Cart removeProductFromCart(Integer customerId, Long productId)
             throws CartException, CustomerException, ProductException;
 
     public Cart removeAllProduct(Integer customerId) throws CartException, CustomerException;
 
-    public Cart increaseProductQuantity(Integer customerId, Integer productId)
+    public Cart increaseProductQuantity(Integer customerId, Long productId)
             throws CartException, CustomerException, ProductException;
 
-    public Cart decreaseProductQuantity(Integer customerId, Integer productId)
+    public Cart decreaseProductQuantity(Integer customerId, Long productId)
             throws CartException, CustomerException, ProductException;
 
 }
